MovementTransferred
Extremely Scuffed Movement but it works:
- camera movement
- walking
- running
- jump (+plus double jump but i'm gonna remove that)
- crouch
- slide
-wall run/jump

This is basically all just moved over from my previous project, I'm going to spend a lot of time tomorrow trying to implement cinemachine and an input manager while also polishing and optimizing a lot of what's already there. But at least for now this should serve as a jump start for some level design testing and other things.